Enable edge to edge support for End of Year
