+++ modules/features/endofyear/src/main/AndroidManifest.xml
@@ -1,4 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
-    <uses-permission android:name="android.permission.DETECT_SCREEN_CAPTURE"/>
+
+    <uses-permission android:name="android.permission.DETECT_SCREEN_CAPTURE" />
+
+    <application>
+        <activity
+            android:name=".StoriesActivity"
+            android:exported="true"
+            android:theme="@style/Theme.EndOfYear" />
+    </application>
 </manifest>
