+++ modules/features/endofyear/src/main/res/values/styles.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <style name="Theme.EndOfYear" parent="android:Theme">
+        <item name="android:windowIsTranslucent">true</item>
+        <item name="android:windowBackground">#80000000</item>
+        <item name="android:windowNoTitle">true</item>
+    </style>
+</resources>
